+++ CHANGELOG.md
@@ -17,6 +17,15 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - Ledger.Rule sub-resource
 - Ledger.Log sub-resource
 - LedgerTransaction resource
+- CreditHolmes resource
+- CreditHolmes.Log sub-resource
+- CreditPreview resource
+- CreditPreview.CreditNotePreview sub-resource
+- CreditNote.Signer.resend_token method
+- CreditNote.pdf method
+- CreditNote.payment method
+- IssuingBillingInvoice resource
+- IssuingBillingTransaction resource
 - IssuingRule missing parameters
 ### Changed
 - IssuingBin resource to IssuingProduct